Destroy supplemental page table in thread_exit()

+++ vm/sup_pte.h
@@ -3,7 +3,7 @@
 
 struct sup_pte
 {
-	uint32_t* vaddr;
+	uint8_t* vaddr;
 	uint64_t access_time;
 	struct hash_elem hash_elem;
 };
